+++ src/main/java/com/forlizzi/medication/entity/user/Medication.java
@@ -7,12 +7,12 @@ import lombok.Data;
 @Data
 @Builder
 public class Medication {
-    private Long medPK;
+    private int medPK;
     private String brandName;
     private boolean hasReaction;
 
     @JsonIgnore
-    public Long getMedPK() {
+    public int getMedPK() {
         return medPK;
     }
